feat: optimize Docker build by installing deps as root before user switch
This change improves Docker layer caching by installing dependencies before creating the non-root user. Dependencies are now installed as root and the non-root user is created afterward to avoid unnecessary rebuilding when application code changes. The ownership of the entire app directory is set to the non-root user to maintain security.